% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vas.R
\encoding{UTF-8}
\name{myfunction}
\alias{myfunction}
\title{Merge the 5 files and gives one in-line CSV and one binded file.}
\usage{
myfunction(d, f1, f2, f3, f4, f5)
}
\arguments{
\item{d}{address of the directory}

\item{f1}{name/address of the biopsy csv}

\item{f2}{name/address of the treatment file}

\item{f3}{name/address of the demographic file}

\item{f4}{name/aadress of the encounter file}

\item{f5}{name/address of the complication file}

\item{address}{function which prmopts the user to add the address of the files.}

\item{myfunction}{function which takes these address as arguement and generates the files as required.}
}
\description{
takes the address of files from the user using prompt and pass these address to another function which cleans and binds the file.The computation time can be high depending upon the size of the file. Be patient.
}
